BigInteger refactoring + Zero division handeling + testing

+++ src/main/java/visitor/Evaluator.java
@@ -4,6 +4,7 @@ import calculator.Expression;
 import calculator.MyNumber;
 import calculator.Operation;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 
 /** Evaluation is a concrete visitor that serves to
@@ -12,13 +13,13 @@ import java.util.ArrayList;
 public class Evaluator extends Visitor {
 
     /** The result of the evaluation will be stored in this private variable */
-    private int computedValue;
+    private BigInteger computedValue;
 
     /** getter method to obtain the result of the evaluation
      *
      * @return an Integer object containing the result of the evaluation
      */
-    public Integer getResult() { return computedValue; }
+    public BigInteger getResult() { return computedValue; }
 
     /** Use the visitor design pattern to visit a number.
      *
@@ -33,14 +34,14 @@ public class Evaluator extends Visitor {
      * @param o The operation being visited
      */
     public void visit(Operation o) {
-        ArrayList<Integer> evaluatedArgs = new ArrayList<>();
+        ArrayList<BigInteger> evaluatedArgs = new ArrayList<>();
         //first loop to recursively evaluate each subexpression
         for(Expression a:o.args) {
             a.accept(this);
             evaluatedArgs.add(computedValue);
         }
         //second loop to accumulate all the evaluated subresults
-        int temp = evaluatedArgs.get(0);
+        BigInteger temp = evaluatedArgs.get(0);
         int max = evaluatedArgs.size();
         for(int counter=1; counter<max; counter++) {
             temp = o.op(temp,evaluatedArgs.get(counter));
